Move study_content files and migrations
