+++ backend-cal/cal_engine/study_content/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class StudyContentConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'cal_engine.study_content'
+
