+++ backend-cal/cal_engine/study_content/signals.py
