--- cal_engine/study_content/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class StudyContentConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'cal_engine.study_content'
-
