--- cal_engine/study_content/serializers.py
@@ -1,13 +1,0 @@
-from rest_framework import serializers
-from .models import Video, Article
-
-class VideoSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Video
-        fields = '__all__'
-
-class ArticleSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Article
-        fields = '__all__'
-
