--- cal_engine/study_content/urls.py
@@ -1,9 +1,0 @@
-from rest_framework.routers import DefaultRouter
-from .views import VideoViewSet, ArticleViewSet
-
-router = DefaultRouter()
-router.register(r'videos', VideoViewSet)
-router.register(r'articles', ArticleViewSet)
-
-
-urlpatterns = router.urls
